requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
ipython==6.2.1
pynubank==0.10
python-dotenv>=0.5.1
